Creation of an annonce view ok, no API request

+++ app/src/main/java/com/esgi/promocare_android/models/annonce/AnnonceDto.kt
@@ -9,7 +9,7 @@ data class AnnonceDto(
     val uuid: String? = null,
     val company_id: String? = null,
     val location: String? = null,
-    val price: Int? = null,
+    val price: Float? = null,
     val promo: Int? = null,
     val status: String? = null,
     val title: String? = null,
