+++ app/src/main/java/com/esgi/promocare_android/models/annonce/AnnonceModel.kt
@@ -4,7 +4,7 @@ data class AnnonceModel(
     val uuid: String? = null,
     val companyId: String? = null,
     val location: String? = null,
-    val price: Int? = null,
+    val price: Float? = null,
     val promo: Int? = null,
     val status: String? = null,
     val title: String? = null,
