+++ app/src/main/java/com/esgi/promocare_android/viewmodel/annonce/CreateAnnonceViewModel.kt
@@ -3,7 +3,7 @@ package com.esgi.promocare_android.viewmodel.annonce
 class CreateAnnonceViewModel {
     var title:String = ""
     var description:String = ""
-    var price:Int = -1
+    var price:Float = -1f
     var category:String = ""
     var location:String = "No location provided"
     var promo:Int = -1
